+++ mobile/android/android-components/components/feature/top-sites/src/main/java/mozilla/components/feature/top/sites/presenter/DefaultTopSitesPresenter.kt
@@ -45,9 +45,9 @@ internal class DefaultTopSitesPresenter(
 
         scope.launch {
             val topSites = storage.getTopSites(
-                innerConfig.totalSites,
-                innerConfig.fetchProvidedTopSites,
-                innerConfig.frecencyConfig
+                totalSites = innerConfig.totalSites,
+                frecencyConfig = innerConfig.frecencyConfig,
+                providerConfig = innerConfig.providerConfig
             )
 
             scope.launch(Dispatchers.Main) {